+++ Main/custom_commands/__init__.py
@@ -1,18 +1,26 @@
 from .command_handlers import setup_commands
 from .views import ImageControlView, LoRAView, OptionsView
 from .web_handlers import handle_generated_image
-from .models import RequestItem
+from .models import RequestItem, ReduxRequestItem, ReduxPromptRequestItem
 from config import CHANNEL_IDS, ALLOWED_SERVERS, BOT_MANAGER_ROLE_ID
-from .database_ops import *
+from .workflow_utils import update_workflow, update_reduxprompt_workflow, validate_workflow
+
 
 __all__ = [
     'setup_commands',
     'ImageControlView',
     'LoRAView', 
     'OptionsView',
+    'ReduxImageView',
+    'ReduxPromptModal',
     'handle_generated_image',
     'RequestItem',
+    'ReduxRequestItem',
+    'ReduxPromptRequestItem',
     'CHANNEL_IDS',
     'ALLOWED_SERVERS',
-    'BOT_MANAGER_ROLE_ID'
+    'BOT_MANAGER_ROLE_ID',
+    'update_workflow',
+    'update_reduxprompt_workflow',
+    'validate_workflow'
 ]